+++ include/klee/Searcher.h
@@ -217,7 +217,7 @@ namespace klee {
     // TODO: use unordered multimap instead
     typedef llvm::SmallPtrSet<ExecutionState*, 8> StatesSet;
     typedef llvm::DenseMap<ExecutionState*, unsigned> StatePosMap;
-    typedef llvm::DenseMap<uint32_t, StatePosMap*> StatesTrace;
+    typedef llvm::DenseMap<uint64_t, StatePosMap*> StatesTrace;
 
     StatesTrace statesTrace;
     StatesSet statesToForward;
